Menu Architecture

+++ Capstone/Class/MainMenu.cs
@@ -37,7 +37,7 @@ namespace Capstone
                         while (!dataInput.EndOfStream)
                         {
                             string[] productArray = dataInput.ReadLine().Split("|");
-                            Product tempProduct = new Product(productArray[1], double.Parse(productArray[2]), productArray[3]);
+                            Product tempProduct = new Product(productArray[1], decimal.Parse(productArray[2]), productArray[3]);
                             Console.WriteLine($"{productArray[1]}|{productArray[2]}");
                         }
                     }
